HSTTWO-1099: Adding full-jcr-query option in query REST operation
svn path=/ecm/site-toolkit/trunk/; revision=22945

+++ site-toolkit/testsuite/sandbox/jaxrs/src/main/java/org/hippoecm/hst/services/support/jaxrs/content/BeanCollection.java
@@ -29,27 +29,27 @@ public class BeanCollection implements Serializable {
     
     private static final long serialVersionUID = 1L;
     
-    private int totalSize = -1;
+    private long totalSize = -1;
     
-    private int beginIndex = -1;
+    private long beginIndex = -1;
     
     public BeanCollection() {
         
     }
     
-    public int getTotalSize() {
+    public long getTotalSize() {
         return totalSize;
     }
     
-    public void setTotalSize(int totalSize) {
+    public void setTotalSize(long totalSize) {
         this.totalSize = totalSize;
     }
     
-    public int getBeginIndex() {
+    public long getBeginIndex() {
         return beginIndex;
     }
     
-    public void setBeginIndex(int beginIndex) {
+    public void setBeginIndex(long beginIndex) {
         this.beginIndex = beginIndex;
     }
